refactor(mcp): close facade core boundaries
